requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
Flask-Mail
Flask-CORS
email-validator
PyJWT
python-dotenv
Werkzeug
requests
gunicorn

